+++ ChangeLog.md
@@ -20,6 +20,9 @@ See docs/process.md for more on how version tagging works.
 
 3.1.15
 ------
+- The JS library helper function `hasExportedFunction` has been replaced with
+  `hasExportedSymbol`, which takes and unmangled (no leading underscore) symbol
+  name.
 - Post-link metadata extraction switched from wasm-emscripten-finalize
   (binaryen) to python code within emscripten.  This change should have no
   semantic effect, but can temporarily be reverted by setting